scripts/requirements.txt
# Requirements file for Python scripts
# Cosine similarity and PDF text extraction

numpy>=1.24.0
scikit-learn>=1.3.0
pdfminer.six>=20221105
python-docx>=1.1.0


